+++ app/src/main/java/cs2263_hw03/Course.java
@@ -1,4 +1,4 @@
-/**
+/*
  * MIT License
  *
  * Copyright (c) 2022 Savannah Webb
@@ -23,11 +23,89 @@
  */
 package cs2263_hw03;
 
-import javafx.application.Application;
-import javafx.scene.Scene;
-import javafx.stage.Stage;
-import java.util.*;
+import java.util.Objects;
+
+/**
+ * creates and hold a course for CourseProcessor
+ * @author Savannah Webb
+ */
 
 public class Course {
+    private String department;
+    private String deptCode;
+    private String courseName;
+    private String courseNumber;
+    private String courseCredits;
+
+    public Course(String department, String deptCode, String courseName, String courseNumber, String courseCredits) {
+        this.department = department;
+        this.deptCode = deptCode;
+        this.courseName = courseName;
+        this.courseNumber = courseNumber;
+        this.courseCredits = courseCredits;
+    }
+
+    public String getDepartment() {
+        return department;
+    }
+
+    public String getDeptCode() {
+        return deptCode;
+    }
+
+    public String getCourseName() {
+        return courseName;
+    }
+
+    public String getCourseNumber() {
+        return courseNumber;
+    }
+
+    public String getCourseCredits() {
+        return courseCredits;
+    }
+
+    public static String[] getDepartmentsArray() {
+        String[] departments = {"Biology", "Chemistry", "Computer Science",
+                "Electrical Engineering", "Mathematics", "Physics"};
+
+        return departments;
+    }
+
+
+    public static String[] courseDept() {
+        String[] deptArray;
+        deptArray = new String[6];
+        deptArray[0] = "CS";
+        deptArray[1] = "MATH";
+        deptArray[2] = "CHEM";
+        deptArray[3] = "PHYS";
+        deptArray[4] = "BIOL";
+        deptArray[5] = "EE";
+
+        return deptArray;
+    }
+
+    public static String courseCode(String deptCode) {
+        String course;
+        if (Objects.equals(deptCode, "Computer Science")) {
+            course = "CS";
+        } else if (Objects.equals(deptCode, "Mathematics")) {
+            course = "MATH";
+        } else if (Objects.equals(deptCode, "Chemistry")) {
+            course = "CHEM";
+        } else if (Objects.equals(deptCode, "Physics")) {
+            course = "PHYS";
+        } else if (Objects.equals(deptCode, "Biology")) {
+            course = "BIOL";
+        } else if (Objects.equals(deptCode, "Electrical Engineering")) {
+            course = "EE";
+        } else {
+            course = "Course Not Found!";
+        }
+
+        return course;
+    }
+
 
 }